Cleanup alimer code base and more vulkan logic.

+++ src/Alimer/Input/Input.cpp
@@ -57,14 +57,14 @@ namespace alimer
         return _mouseButtons.IsHeld(static_cast<uint32_t>(button));
     }
 
-    void Input::MouseButtonEvent(MouseButton button, int32_t x, int32_t y, bool pressed)
+    void Input::MouseButtonEvent(MouseButton button, float x, float y, bool pressed)
     {
         _mouseButtons.Post(static_cast<uint32_t>(button), pressed);
         _mousePosition.x = x;
         _mousePosition.y = y;
     }
 
-    void Input::MouseMoveEvent(MouseButton button, int32_t x, int32_t y)
+    void Input::MouseMoveEvent(MouseButton button, float x, float y)
     {
         ALIMER_UNUSED(button);
         _mousePosition.x = x;
